WIP
This commit adds an async function both for getAllTrips and for getSingleTraveler in the travel_tracker_service.js file. Adds the traveler_dashboard.js file, and inside this file are the functions getTravelerInfo and getTripsArr which both trigger their corresponding ASYNC functions in the travel_tracker_service file. These are both set to console.log the data retrieved from the functions.