Add traveler dashboard
This commit adds the traveler dashboard. After logging in, the
travelers' ID number is used to fetch the respective data for that
traveler so that the issues for this iteration can begin to be
completed.

In order to have the travelers' data, multiple asynchronous
functions fetch and return it. In order to show the
data correctly on the page, the data is rendered on the page via
helper functions. Helper functions were created to eliminate repititous
code, and help to better organize the project. This completes the
iteration of the project as the total amount the traveler has spent,
the travelers' upcoming trips, and the travelers' past trips are all
showing on the page. The page is styled with SCSS and is mobile
responsive.

Resolves: #13