WIP
I have started a function called matchTravToTrip which I am trying to get to match the traveler ID which is taken from the input and match it with the <i>userID</i> from the trips array. So far any attempt I have made to link them up has been unsucessful. I am seeing the UserID as 6 and I am inputting that as my Traveler ID when logging in...I have verified they are both set to numbers and can see they are both 6. When I verify 6 === 6 in the console this is coming back as true...But when I try and verify the IDS are matching via an IF statement I am having no luck. Will sleep on it and continue research / development tomorrow.