Populate trip data and get trip cost info
This commit adds JS to populate the travelers trip data on the screen. Still needs to be styled. This commit also checks the destination ID's from the traveler trip info with the destination ID's from the destination array. As of now the itemized costs are being console.log'd. Will continue working on functionality to get the total cost information + agent fee into an array, and also stylyzing the traveler trip information.