Complete initial styling of traveler dashboard
This commit adds scss styling to the traveler dashboard. Adds border
to both upcoming and previous trips sections. If there are more than 3
trip / destination results there is a sliding function allowing
them to be seen while still allowing the general flow and styling of
the page to not be interrupted.