Add all traveler interactivity functionality
This commit will add the option for the traveler to make a trip request. The traveler will search for a destination, choose the number of travelers, and choose a start and end date. Once this information is selected, the traveler will click the 'CALCULATE' button to calculate and then populate the trip request information. After the cost and trip information is shown, the traveler has the option to 'CONFIRM' the trip. Once the trip has been confirmed, the trip is then added to the TRIPS endpoint, and that requested trip will be shown as 'PENDING' in that travelers 'UPCOMING TRIPS' section. Once the agent logs in and approves the trip, the trip cost is added to the travelers total amount spent, and the trip will show as APPROVED in the travelers 'UPCOMING TRIPS' section.

There were multiple instances of me using arrays to handle data which resulted in me later having to hard code the position of the data I needed in order to use it. This commit corrects these instances and replaces the arrays with objects, which allow me to simply call the data from the object instead of hardcoding its position in the array.
This commit also refactors any function that is longer than 10 lines into smaller more manageable helper functions.

All functions have been re-organized to be in their logical place in terms of their relationship to either the general dashboard functionality, the agent dashboard functionality, or the traveler dashboard functionality.
I was able to split up most the larger functions into smaller helper functions, but there are instances where I thought the result would be redundant or take up as much space refactoring it as it took up originally so those have been left.
These are:

in dashboard_helper.js:
filterPastAndUpcomingTrips()
calculateAndReturnCost()
calculateTripRequestCost()
prepareTripRequestData()
renderTripRequest()
returnTripRequestData()

in agent_dashboard_helper.js:
returnApprovedTripsFlightsAndLodging() getLodgingCostsForAgent()

in traveler_dashboard_helper.js:
makePastAndUpcomingTripsArrays() prepareTravelerDashboardData()