Add POST request functionality
This commit adds the functionality to the CONFIRM TRIP button. When the user clicks to confirm the trip, the trip is added to the travelers "pending trips". As of right now you have to reload the page and login to see the new pending trip in the "upcoming trips" section on their dashboard, eventually this will be done dynamically without the need to refresh