Add agent interactivity
This commit allows the agent to approve / keep as pending any pending trips. The agent also has the option to search for a traveler by name. Once the name has been selected and the 'SHOW TRAVELER' button clicked, the travelers trips and total amount spent. If the traveler has any 'PENDING' trips, there will be an option for the agent to either approve or keep the trip as pending.