Add agent interactivity
This commit adds the agent interactivity functionality to the project. With this commit the agent will be able to approve any of the trips in the "REQUESTED TRIPS" section. The agent will also have the ability to search for a traveler by name and view all their trips and then approve / delete pending trips.