Add APPROVE / KEEP AS PENDING buttons
This commit adds the buttons and button functionality for the agent to approve trips.
The agent will be able to approve the trips in the 'REQUESTED TRIPS' section. At this time they will not be able to approve / keep as pending trips in the TODAYS TRIPS section, although if there is a PENDING trip showing for today, the agent will be able to approve that in the requested trips section.