Update page styling and responsiveness
This commit updates the styling for the project. In order to make the page look more professional I have added a header with a randomized image displaying each time the page is refreshed. The destinations that appear below the header are also randomized. The project is fully responsive as far as I can tell at this time. Upon further testing I will address any further issues.
On the agent dashboard page, now when there are no trips on the current date there is a message saying "No travelers on trips today".

This commit also adds the 'Logout' button. This will show underneath the Traveler / Agent name div.