Correct responsiveness on iphone XR
This commit corrects the responsiveness issues on Iphone XR:
When logging in as agent the APPROVE trip button will now have the correct padding.

When logging in as traveler:
The trip request form will now have the correct padding and not be as jumbled.
The CALCULATE button on the trip request form will now have the correct amount of padding.